feat (PlayerName) : 원시값 name을 PlayerName으로 포장 객체 구현
